+++ NEWS.md
@@ -1,8 +1,13 @@
 # motus 4.0.0
+
 ### Small Changes
 * Add `test` to metadata in `tagDeps` table (`tagDeployTest` in `alltags` and
   `alltagsGPS` views)
 * Add `age` and `sex` to metadata in `tagDeps` table
+* For CTT V2
+    * Add `lat_mean`, `lon_mean`, and `n_fixes` to `gps` table
+    * Add `nodets`, `firmware`, `solarVolt`, `solarCurrent`, `solarCurrentCumul`, `lat`, and `lon` to `nodeData` table
+    * Add `validated` to `hits` table
 
 # motus 3.0.1